+++ src/bgp/message/mod.rs
@@ -3,7 +3,7 @@ pub mod update;
 pub mod nlri;
 pub mod notification;
 pub mod keepalive;
-pub mod attribute_list;
+pub mod attribute;
 
 use octseq::{Octets, Parser};
 use crate::util::parser::ParseError;
