+++ src/bgp/mod.rs
@@ -1,5 +1,6 @@
 pub mod communities;
 pub mod types;
+pub mod route;
 
 pub mod message;
 pub use crate::util::parser::ParseError;
